+++ config.py
@@ -1,10 +1,10 @@
 # Single-sourced project configuration values:
 
 # the full version, including alpha/beta/rc tags
-release = '0.1.0'
+release = '0.2.0'
 
 # the short X.Y version
-version = '0.1'
+version = '0.2'
 
 # distribution package name
 name = 'wpyprint'