+++ tests/ppcontext/test_b_format.py
@@ -1,4 +1,4 @@
-# test_a_format
+# test_b_format
 
 from wpyprint import PPContext
 
@@ -6,33 +6,41 @@ from wpyprint import PPContext
 def test_empty_1():
     ppc = PPContext()
     result = ppc.format()
-    # print(result)
-    assert result == "--"
+    # print("\n" + result)
+    assert result == ""
+
+    result = ppc.format("")
+    # print("\n" + result)
+    assert result == ''
 
 
 def test_str_1():
     ppc = PPContext()
-    result = ppc.format("")
-    # print(result)
-    assert result == '""'
+    result = ppc.format("foobar")
+    # print("\n" + result)
+    assert result == 'foobar'
+
+    result = ppc.format("foo\nbar")
+    # print("\n" + result)
+    assert result == 'foo\nbar'
 
 
 def test_list_1():
     val = []
     ppc = PPContext()
     result = ppc.format(val)
-    # print(result)
+    # print("\n" + result)
     assert result == """[]"""
 
     val = ['abc', 'def', 'ghi', 'jkl']
     ppc = PPContext()
     result = ppc.format(val)
-    # print(result)
+    # print("\n" + result)
     assert result == """[abc, def, ghi, jkl]"""
 
     ppc = PPContext(width=16)
     result = ppc.format(val)
-    # print(result)
+    # print("\n" + result)
     assert result == """- abc
 - def
 - ghi
@@ -43,56 +51,66 @@ def test_dict_1():
     val = {}
     ppc = PPContext()
     result = ppc.format(val)
-    # print(result)
+    # print("\n" + result)
     assert result == """{}"""
 
+
+def test_dict_2():
     val = {'abc': 123, 'def': 456, 'ghi': 789}
     ppc = PPContext()
     result = ppc.format(val)
-    # print(result)
-    assert result == """abc: 123
-def: 456
-ghi: 789"""
+    # print("\n" + result)
+    assert result == """- abc: 123
+- def: 456
+- ghi: 789"""
 
+
+def test_dict_3():
     val = {'abc': 123, 'def': 456, 'ghi': 789}
     ppc = PPContext(width=16)
     result = ppc.format(val)
-    # print(result)
-    assert result == """abc: 123
-def: 456
-ghi: 789"""
+    # print("\n" + result)
+    assert result == """- abc: 123
+- def: 456
+- ghi: 789"""
 
+
+def test_dict_4():
     val = {'abcdefghi': 123456789, 'def': 456, 'ghi': 789}
     ppc = PPContext(width=16)
     result = ppc.format(val)
-    # print(result)
-    assert result == """abcdefghi:
-  123456789
-def: 456
-ghi: 789"""
+    # print("\n" + result)
+    assert result == """- abcdefghi:
+    123456789
+- def: 456
+- ghi: 789"""
 
+
+def test_dict_5():
     val = {
         'abcdefghi': "abc def ghi jkl mno pqr stu",
         'def': 456,
         'ghi': 789,
     }
-    ppc = PPContext(width=16)
+    ppc = PPContext(width=18)
     result = ppc.format(val)
-    # print(result)
-    assert result == """abcdefghi: abc
-  def ghi jkl
-  mno pqr stu
-def: 456
-ghi: 789"""
+    # print("\n" + result)
+    assert result == """- abcdefghi: abc
+    def ghi jkl
+    mno pqr stu
+- def: 456
+- ghi: 789"""
 
+
+def test_dict_6():
     val = {'a': 1, 'b': 2, 'c': 3, 'd': 4, 'e': 5, 'f': 6, 'g': 7, 'h': 8}
     ppc = PPContext(truncate=4)
     result = ppc.format(val)
-    # print(result)
-    assert result == """a: 1
-b: 2
-c: 3
-d: 4"""
+    # print("\n" + result)
+    assert result == """- a: 1
+- b: 2
+- c: 3
+- d: 4"""
 
 
 def test_composite_1():
@@ -102,10 +120,10 @@ def test_composite_1():
     ]
     ppc = PPContext()
     result = ppc.format(val)
-    # print(result)
-    assert result == """- abc: 123
-  def: 456
-- ghi: 789"""
+    # print("\n" + result)
+    assert result == """- - abc: 123
+  - def: 456
+- - ghi: 789"""
 
 
 def test_composite_2():
@@ -114,18 +132,18 @@ def test_composite_2():
         'abcdefghi': "abc def ghi jkl mno pqr stu",
         'ghi': {'a': 1, 'b': 2, 'c': 3, 'd': 4, 'e': 5, 'f': 6, 'g': 7},
     }
-    ppc = PPContext(width=16, truncate=4)
+    ppc = PPContext(width=18, truncate=4)
     result = ppc.format(val)
-    # print(result)
-    assert result == """abc: [123, 456]
-abcdefghi: abc
-  def ghi jkl
-  mno pqr stu
-ghi:
-  a: 1
-  b: 2
-  c: 3
-  d: 4"""
+    # print("\n" + result)
+    assert result == """- abc: [123, 456]
+- abcdefghi: abc
+    def ghi jkl
+    mno pqr stu
+- ghi:
+    - a: 1
+    - b: 2
+    - c: 3
+    - d: 4"""
 
 
 def test_composite_3():
@@ -134,5 +152,11 @@ def test_composite_3():
     }
     ppc = PPContext(width=16, truncate=4)
     result = ppc.format(val)
-    # print(result)
-    assert result == """abc: [1, 2, 3]"""
+    # print("\n" + result)
+    assert result == """- abc: [1, 2, 3]"""
+
+# def test_bullet_1():
+#     ppc = PPContext()
+#     result = ppc.format(val)
+#     # print("\n" + result)
+#     assert result == """abc: [1, 2, 3]"""