+++ wpyprint/utils/__init__.py
@@ -1,1 +1,1 @@
-from .predicates import is_dict, is_multiline, is_set, is_singleline
+from .predicates import is_dict, is_multiliner, is_oneliner, is_set, is_tuple