+++ CHANGELOG.rst
@@ -4,3 +4,8 @@ Changelog
 v0.1 - 2019-01-02
 +++++++++++++++++
 - Added the 'PPContext' class and the 'print' function.
+
+v0.2 - 2019-01-06
++++++++++++++++++
+- Name-value pairs are now also bulletted.
+- Added optional 'bullet' parameter where appropriated.